+++ pylabrobot/liquid_handling/backends/tecan/__init__.py
@@ -1,1 +1,1 @@
-from .EVO import EVOBackend
+from .EVO_backend import EVOBackend
